Remove principal from SettingsContext since it depends on the particular auth used
