+++ packages/react-components/src/contexts/SettingsContext.tsx
@@ -16,7 +16,6 @@ import {
   Streams,
   genRunID,
 } from '@runmedev/react-console'
-import { JwtPayload, jwtDecode } from 'jwt-decode'
 import { Subscription } from 'rxjs'
 import { ulid } from 'ulid'
 
@@ -30,7 +29,6 @@ interface Settings {
 }
 
 interface SettingsContextType {
-  principal: string
   checkRunnerAuth: () => void
   createAuthInterceptors: (redirect: boolean) => Interceptor[]
   defaultSettings: Settings
@@ -71,21 +69,6 @@ export const SettingsProvider = ({
   createAuthInterceptors,
 }: SettingsProviderProps) => {
   const [runnerError, setRunnerError] = useState<StreamError | null>(null)
-
-  const principal = useMemo(() => {
-    const token = getSessionToken()
-    if (!token) {
-      return 'unauthenticated'
-    }
-    let decodedToken: JwtPayload & { email?: string }
-    try {
-      decodedToken = jwtDecode(token)
-      return decodedToken.email || decodedToken.sub || 'unauthenticated'
-    } catch (e) {
-      console.error('Error decoding token', e)
-      return 'unauthenticated'
-    }
-  }, [])
 
   const defaultSettings: Settings = useMemo(() => {
     const isLocalhost = window.location.hostname === 'localhost'
@@ -237,7 +220,6 @@ export const SettingsProvider = ({
   return (
     <SettingsContext.Provider
       value={{
-        principal,
         checkRunnerAuth,
         createAuthInterceptors: actualCreateAuthInterceptors,
         defaultSettings,
